% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dooneboot.R
\name{dOneBoot}
\alias{dOneBoot}
\alias{doOneBoot}
\title{Fonction dOneBoot}
\usage{
doOneBoot(i)
}
\arguments{
\item{i}{la ligne i correspondant aux paramètres dans boot_params}
}
\value{
coefficient et la matrice de variance covariance
}
\description{
Fonction dOneBoot
}
