% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dooneboot.R
\name{doOneBoot}
\alias{doOneBoot}
\title{Fonction dOneBoot}
\usage{
doOneBoot(boot_params, i)
}
\arguments{
\item{boot_params}{fichier avec l'ensemble des paramètres bootstrap}

\item{i}{la ligne i correspondant aux paramètres dans boot_params}
}
\value{
coefficient et la matrice de variance covariance
}
\description{
Fonction dOneBoot
}
