% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doonebootWCIE.R
\name{doOneBootWCIE}
\alias{doOneBootWCIE}
\title{Bootstrap Estimation of WCIE for a Single Replicate}
\usage{
doOneBootWCIE(
  boot_params,
  i,
  times,
  mexpo,
  var.time,
  weightbasis,
  knots,
  knots.vector,
  data,
  reg.type,
  model
)
}
\arguments{
\item{boot_params}{A matrix or data frame containing all bootstrap parameter vectors.}

\item{i}{Integer. The row index corresponding to a bootstrap sample in \code{boot_params}.}

\item{mexpo}{An object of class \code{hlme} from the \code{lcmm} package, used to model the exposure process.
The object must be created using the \code{hlme} function with specific arguments provided.
The fixed effects formula and the random effects formula must be specified. The \code{subject} argument
must indicate the subject ID, and the dataset must be provided via the \code{data} argument.
It is essential to include \code{returnData = TRUE} in the function call to ensure that the internal data can be accessed.}

\item{var.time}{character indicating the name of the time variable
in the model \code{mexpo}.}

\item{weightbasis}{Type of temporal weighting function used to estimate the Weighted Cumulative Indirect Effects (WCIE).
This specifies the functional form used to model the influence of past exposures over time.
Currently, the following options are available: \code{"NS"} for natural splines (implemented),
\code{"BS"} for B-splines (to be developed), and \code{"PS"} for P-splines (to be developed).}

\item{knots}{number of internal knots}

\item{knots.vector}{Vector of internal knots for the splines (used only for splines temporal weighting function).}

\item{data}{A data frame containing the variables specified in the outcome model
\code{model} including the outcome variable. This dataset will be used to estimate the
outcome model, and the WCIE variables calculated previously will be added to this
dataset prior to model fitting.}

\item{reg.type}{Type of outcome model: \code{"logistic"}, \code{"cox"}, ...}

\item{model}{two-sided linear formula object for the outcome model. The response outcome is on
the left of ~ and the covariates are separated by + on the right of ~. To include the effect of past exposure,
you must explicitly add \code{WCIE} (or interaction terms such as \code{WCIE:sex}) to the formula.
For example, \code{Y ~ WCIE + age + sex} or \code{Y ~ WCIE:sex + age} are valid formulas.}

\item{timerange}{Numeric vector of length 2
indicating the desired time window for exposure (min, max).}

\item{step}{Step between two consecutive time points in the time window indicating in \code{timerange}.}
}
\value{
A list containing:
  \item{coef}{Estimated coefficients of the outcome model.}
  \item{V}{Variance-covariance matrix (intra + inter) of the estimators.}
  \item{AIC}{AIC criteria of the outcome model.}
  \item{loglike}{log-likelihood of the outcome model.}
}
\description{
This internal function is used to perform one iteration of the bootstrap procedure
for estimating WCIE and fitting the corresponding outcome model using perturbed parameters.
}
\details{
The function takes a set of bootstrap parameters from \code{boot_params}, reconstructs the exposure model
using those parameters, and then estimates the WCIE values. These WCIE components are then used to
fit the outcome model. It returns the estimated regression coefficients, their variance-covariance matrix,
the AIC, and the log-likelihood of the outcome model.
}
