% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.WCEland.R
\name{plot.WCEland}
\alias{plot.WCEland}
\title{summary.WCIE2F}
\usage{
\method{plot}{WCEland}(x, ...)
}
\arguments{
\item{x}{An object of class \code{WCEland} returned by the \code{WCEland} function.}

\item{...}{Further arguments (currently not used).}
}
\value{
Prints a plot of the exposition historic effects estimates.
}
\description{
summary.WCIE2F
}
