% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.WCIE2F.R
\name{plot.WCIE2F}
\alias{plot.WCIE2F}
\title{summary.WCIE2F}
\usage{
\method{plot}{WCIE2F}(object, ...)
}
\arguments{
\item{object}{An object of class \code{WCIE2F} returned by the \code{WCIE2F} function.}

\item{...}{Further arguments (currently not used).}
}
\value{
Prints a plot of the exposition historic effects estimates.
}
\description{
summary.WCIE2F
}
\examples{
\dontrun{
}
}
