% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWCIE.R
\name{simulateWCIE}
\alias{simulateWCIE}
\title{Data simulation according to different outcome models from WCIEpackage}
\usage{
simulateWCIE(
  object,
  nsim = 1,
  seed = NULL,
  times,
  internal.step,
  tname,
  n,
  Xbin = NULL,
  Xcont = NULL,
  weightbasis,
  reg.type,
  knots = NULL,
  knots.vector = NULL,
  coef.wcie,
  Xoutcome
)
}
\arguments{
\item{object}{an object of class \code{hlme}}

\item{nsim}{simulation number}

\item{seed}{the random seed}

\item{times}{either a data frame with 2 columns containing IDs and measurement times, or a
vector of length 4 specifying the minimal and maximum measurement times, the spacing between
2 consecutive visits and the margin around this spacing}

\item{internal.step}{Time increment used for high-resolution simulation before resampling to the user's desired time step.}

\item{tname}{the name of the variable representing the measurement times in \code{object}. Default
to the second column's name of times if it is a data frame, and to object$var.time otherwise.}

\item{n}{number of subjects to simulate*}

\item{Xbin}{an optional named list giving the probabilities of the binary covariates to
simulate. The list's names should match the binary covariate's names used in \code{object}.}

\item{Xcont}{an optional named list giving the mean and standard deviation of the Gaussian
covariates to simulate. The list's names should match the continuous covariate's names used
in object.}

\item{weightbasis}{Type of temporal weighting function used to estimate the Weighted Cumulative Indirect Effects (WCIE).
This specifies the functional form used to model the influence of past exposures over time.
Currently, the following options are available: \code{"NS"} for natural splines (implemented),
\code{"BS"} for B-splines (to be developed), and \code{"PS"} for P-splines (to be developed).}

\item{reg.type}{Type of outcome model: \code{"logistic"}, \code{"cox"}, ...}

\item{knots}{number of internal knots}

\item{knots.vector}{Vector of internal knots for the splines (used only for splines temporal weighting function).}

\item{coef.wcie}{named list giving the parameters of the WCIE variables to
simulate.}

\item{Xoutcome}{named list giving the parameters of the intercept and the covariables in the
logistic model to simulate de outcome variable. The list's names should match the intercept and
covariable names used in object}
}
\value{
\item{exposition.data}{a data frame with one line per observation and one column per variable.
 Variables appears in the following order : subject id, measurement time, entry time, binary
 covariates, continuous covariates, longitudinal outcomes.}
\item{outcome.data}{a data frame with one line per individual and one column per variable.
 Variables appears in the following order : subject id, binary
 covariates, continuous covariates, binary outcomes.}
}
\description{
This function simulates two sample, one exposition sample according to a model estimated
with hlme, functions and an outcome sample simulate in relation with the effect of the exposition.
}
\details{
to do
}
