% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.WCEland.R
\name{summary.WCEland}
\alias{summary.WCEland}
\title{summary.WCEland}
\usage{
\method{summary}{WCEland}(object, ...)
}
\arguments{
\item{object}{An object of class \code{WCEland} returned by the \code{WCEland} function.}

\item{...}{Further arguments (currently not used).}
}
\value{
Prints a summary of the WCEland analysis including model information, estimation results, and bootstrap details.
}
\description{
summary.WCEland
}
\examples{
\dontrun{
res <- WCEland(mexpo = mexpo,
    var.time = "TIME",
    timerange = c(0, 10),
    data = data,
    model = outcome_model)
summary(res)
}
}
