% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.WCIE2F.R
\name{summary.WCIE2F}
\alias{summary.WCIE2F}
\title{summary.WCIE2F}
\usage{
\method{summary}{WCIE2F}(object, ...)
}
\arguments{
\item{object}{An object of class \code{WCIE2F} returned by the \code{WCIE2F} function.}

\item{...}{Further arguments (currently not used).}
}
\value{
Prints a summary of the WCIE2F analysis including model information, estimation results, and bootstrap details.
}
\description{
summary.WCIE2F
}
\examples{
\dontrun{
res <- WCIE2F(mexpo = mexpo,
    var.time = "TIME",
    timerange = c(0, 10),
    data = data,
    model = outcome_model)
summary(res)
}
}
