% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcie_estimation.R
\name{WCIE_estimation}
\alias{WCIE_estimation}
\title{Fonction WCIE_estimation}
\usage{
WCIE_estimation(
  mexpo,
  var.time,
  timerange,
  step,
  weightbasis,
  knots,
  data,
  reg.type,
  model
)
}
\arguments{
\item{mexpo}{l'objet hlme contenant le modèle de toute l'exposition}

\item{var.time}{la variable temporelle utilisée}

\item{timerange}{fenêtre de temps d'exposition que l'utilisateur souhaite analyser}

\item{step}{pas de temps entre chaque observation dans la fenêtre de temps}

\item{weightbasis}{la fonction à attribuer au poids de chaque temps d"exposition}

\item{knots}{le nombre de noeuds internes si utilisation de splines}

\item{data}{les données avec la variable à étudier}

\item{reg.type}{le type de regression à utiliser pour étudier la variable d'intérêt}

\item{model}{la formule du modèle à utiliser pour étudier la variable}
}
\value{
estimation du modèle avec l'outcome
}
\description{
Fonction WCIE_estimation
}
