% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCIE2F.R
\name{WCIE2F}
\alias{WCIE2F}
\title{Fonction WCIE2F}
\usage{
WCIE2F(
  mexpo,
  var.time,
  timerange,
  step = 1,
  weightbasis = "NS",
  knots = 3,
  data,
  reg.type = "RL",
  model,
  n_boot = 500
)
}
\arguments{
\item{mexpo}{l'objet hlme contenant le modèle de toute l'exposition}

\item{var.time}{la variable temporelle utilisée}

\item{timerange}{fenêtre de temps d'exposition que l'utilisateur souhaite analyser}

\item{step}{pas de temps entre chaque observation dans la fenêtre de temps}

\item{weightbasis}{la fonction à attribuer au poids de chaque temps d"exposition}

\item{knots}{le nombre de noeuds internes si utilisation de splines}

\item{data}{les données avec la variable à étudier}

\item{reg.type}{le type de regression à utiliser pour étudier la variable d'intérêt}

\item{model}{la formule du modèle à utiliser pour étudier la variable}

\item{n_boot}{nombre d'échantillon bootsrap à tirer pour le calcul de la variance}
}
\value{
estimation du modèle avec l'outcome + variance calculer avec bootstrap paramétrique
}
\description{
Fonction WCIE2F
}
